- Fixed some warnings. - mforms: added new node functions to manipulate existing child nodes in a treenode. Win + TUT implementation added too. - Win: mforms: switched off LoadOnDemand in TreeNodeAdv. This doesn't affect backend loading (e.g. in the schema tree) but avoids things like the expand icon to appear always for new or moved nodes, regardless whether they have child nodes or not. - grt: added ListRef move() call for List::move(). - Win: ANTLR parser: extended conditions to regenerate the parser files to the new simple parser. - trigger editor: added move up + down to the context menu for the trigger list.
[TESTS]
- Checked that treeviews (especially the schema tree) is not affected by switching off LoadOnDemand.
- Checked that triggers can be moved with the context menu, including to new groups, automatically changing their timing and event.

+++ library/forms/winforms/src/wf_textentry.cpp
@@ -306,7 +306,7 @@ void TextEntryWrapper::paste(mforms::TextEntry *self)
 void TextEntryWrapper::select(mforms::TextEntry *self, const base::Range &range)
 {
   TextBox ^textbox = TextEntryWrapper::GetManagedObject<TextBox>(self);
-  textbox->Select(range.position, range.size == (size_t)-1 ? textbox->Text->Length : range.size);
+  textbox->Select((int)range.position, range.size == (size_t)-1 ? textbox->Text->Length : (int)range.size);
 }
 
 //--------------------------------------------------------------------------------------------------
